fix: resolve gosec security warnings
- Fix integer overflow conversion warnings (G115) by adding range validation
- Fix implicit memory aliasing in for loop (G601) using index-based access
- Add bounds checking before int to uint32 conversions
- Improve Bitcoin protocol safety and security compliance

+++ pkg/bitcoin/script.go
@@ -594,13 +594,13 @@ func (se *ScriptEngine) bytesToNum(data []byte) int64 {
 			// Last byte: check sign bit
 			if data[i]&0x80 != 0 {
 				// Negative number
-				result |= int64(data[i]&0x7f) << (8 * uint(i))
+				result |= int64(data[i]&0x7f) << (8 * uint64(i))
 				result = -result
 			} else {
-				result |= int64(data[i]) << (8 * uint(i))
+				result |= int64(data[i]) << (8 * uint64(i))
 			}
 		} else {
-			result |= int64(data[i]) << (8 * uint(i))
+			result |= int64(data[i]) << (8 * uint64(i))
 		}
 	}
